[Triton][Kernel] Refactor: organize triton linear kernels into layers/triton_ops
Move Triton kernel mm_triton() and addmm_triton()
to layers/triton_ops/linear.py
